refactor: extract renderer-agnostic AdminModel from AdminRouter
Foundation for go/ts conformance parity (the root-cause fix for the
O–S cascade: only fastapi has behavioral conformance in CI). The
test-admin router's spec-derived logic — state-field -> {pk: row}
projection inference, primary-key resolution, datetime-field
detection — is language-agnostic; only its rendering was
Python/SQLAlchemy-coupled.

Moved that logic verbatim into a new `AdminModel` object;
`AdminRouter` now consumes it and renders the identical
Python/FastAPI output (fastapi goldens byte-identical —
EmitPythonTest 3/3). `AdminRouter` keeps thin `private[testgen]`
forwarders so existing Behavioral.scala callers are untouched. No
behavior change; `supportsTestgen` unchanged (ts/go still rejected).
A ts-express renderer over the same model + CI conformance wiring
follows.
